refactored router to previous style. Commented test files
