+++ src/tests/itemCategory.test.ts
@@ -1,11 +1,11 @@
-import request from "supertest";
+// import request from "supertest";
 
-describe("GET /itemCategories", () => {
-  it("responds with a list of itemCategories", async () => {
-    const response = await request(
-      "https://clumsy-gray-angler.cyclic.app/"
-    ).get("/itemCategory/get");
-    expect(response.status).toBe(200);
-    // expect(response.body).toHaveLength(48);
-  });
-});
+// describe("GET /itemCategories", () => {
+//   it("responds with a list of itemCategories", async () => {
+//     const response = await request(
+//       "https://clumsy-gray-angler.cyclic.app/"
+//     ).get("/itemCategory/get");
+//     expect(response.status).toBe(200);
+//     // expect(response.body).toHaveLength(48);
+//   });
+// });
